Fix the known runtime errors pointed out by ubsan
Summary:
As preparation for enabling a dedicated UndefinedBehaviorSanitizer (`ubsan`) build we need to fix all the known issues exposed by pre-commit MySQL tests so that we can have a clean baseline.

Fixes here are trivial: fix a casting issue and make sure that we'll shift the bits in the allowed range only. There are tricker (and more indecipherable) ways to fix them, but I chose readability over brevity.

Test Plan:
  # Produce a `ubsan` build.
  # Run `mysqltest.sh` and `mysqltest.sh --testset=RocksDB`.
  # Make sure that all the test cases pass.

Reviewers: MarkCallaghan, vasilep

Reviewed By: vasilep

Subscribers: jkedgar, webscalesql-eng

Differential Revision: https://reviews.facebook.net/D55779

+++ sql-common/client.c
@@ -4976,7 +4976,7 @@ csm_parse_handshake(mysql_csm_context *ctx)
     /* New protocol with 16 bytes to describe server characteristics */
     mysql->server_language=end[2];
     mysql->server_status=uint2korr(end+3);
-    mysql->server_capabilities|= uint2korr(end+5) << 16;
+    mysql->server_capabilities |= ((unsigned long) uint2korr(end + 5) << 16);
     ctx->pkt_scramble_len= end[7];
     if (ctx->pkt_scramble_len < 0)
     {
